# docs/public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pricore API",
        "version": "v1",
        "description": "REST API for managing a Pricore registry: organizations, repositories, packages, members, mirrors, and access tokens. Authenticate with a personal or organization access token sent as `Authorization: Bearer <token>`."
    },
    "servers": [
        {
            "url": "http://pricore.test/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/organizations/{organization}/invitations/{invitation}": {
            "delete": {
                "operationId": "v1.invitations.destroy",
                "tags": [
                    "Invitation"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "invitation",
                        "in": "path",
                        "required": true,
                        "description": "The invitation UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/organizations/{organization}/members/{member}": {
            "patch": {
                "operationId": "v1.members.update",
                "tags": [
                    "Member"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMemberRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`OrganizationMemberData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationMemberData"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.members.destroy",
                "tags": [
                    "Member"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "description": "An error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview.",
                                            "example": "Cannot remove the organization owner."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/organizations/{organization}/mirrors/{mirror}": {
            "get": {
                "operationId": "v1.mirrors.show",
                "tags": [
                    "Mirror"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mirror",
                        "in": "path",
                        "required": true,
                        "description": "The mirror UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`MirrorData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MirrorData"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.mirrors.destroy",
                "tags": [
                    "Mirror"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mirror",
                        "in": "path",
                        "required": true,
                        "description": "The mirror UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/organizations/{organization}": {
            "get": {
                "operationId": "v1.organizations.show",
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`OrganizationData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationData"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "patch": {
                "operationId": "v1.organizations.update",
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrganizationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`OrganizationData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationData"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.organizations.destroy",
                "tags": [
                    "Organization"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/organizations/{organization}/packages/{package}": {
            "get": {
                "operationId": "v1.packages.show",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PackageData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PackageData"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.packages.destroy",
                "tags": [
                    "Package"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/organizations/{organization}/packages/{package}/versions/{version}": {
            "delete": {
                "operationId": "v1.packages.versions.destroy",
                "tags": [
                    "PackageVersion"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "package",
                        "in": "path",
                        "required": true,
                        "description": "The package UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "description": "The version UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/organizations/{organization}/repositories/{repository}": {
            "get": {
                "operationId": "v1.repositories.show",
                "tags": [
                    "Repository"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "repository",
                        "in": "path",
                        "required": true,
                        "description": "The repository UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`RepositoryData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RepositoryData"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.repositories.destroy",
                "tags": [
                    "Repository"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "repository",
                        "in": "path",
                        "required": true,
                        "description": "The repository UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/organizations/{organization}/tokens/{token}": {
            "patch": {
                "operationId": "v1.tokens.update",
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "description": "The token UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAccessTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AccessTokenData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessTokenData"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.tokens.destroy",
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "description": "The organization slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "description": "The token UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/user/tokens/{token}": {
            "patch": {
                "operationId": "v1.user.tokens.update",
                "tags": [
                    "UserToken"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "description": "The token UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAccessTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AccessTokenData`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessTokenData"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.user.tokens.destroy",
                "tags": [
                    "UserToken"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "description": "The token UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "AccessTokenData": {
                "type": "object",
                "additionalProperties": {},
                "title": "AccessTokenData"
            },
...
            "MirrorData": {
                "type": "object",
                "additionalProperties": {},
                "title": "MirrorData"
            },
            "OrganizationData": {
                "type": "object",
                "additionalProperties": {},
                "title": "OrganizationData"
            },
...
            "OrganizationMemberData": {
                "type": "object",
                "additionalProperties": {},
                "title": "OrganizationMemberData"
            },
            "PackageData": {
                "type": "object",
                "additionalProperties": {},
                "title": "PackageData"
            },
...
            "RepositoryData": {
                "type": "object",
                "additionalProperties": {},
                "title": "RepositoryData"
            },
...
            "UpdateAccessTokenRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "scopes": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "$ref": "#/components/schemas/TokenScope"
                        }
                    }
                },
                "required": [
                    "name"
                ],
                "title": "UpdateAccessTokenRequest"
            },
            "UpdateMemberRoleRequest": {
                "type": "object",
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "member"
                        ]
                    }
                },
                "required": [
                    "role"
                ],
                "title": "UpdateMemberRoleRequest"
            },
            "UpdateOrganizationRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "required": [
                    "name"
                ],
                "title": "UpdateOrganizationRequest"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}